replaced findOrFail with find

+++ app/Http/Controllers/ImageController.php
@@ -13,7 +13,10 @@ class ImageController extends Controller
     }
 
     public function destroy(string $id){
-        $data = Image::findOrFail($id);
+        $data = Image::find($id);
+        if(! $data){
+            return response()->json(["message" => "Target with id={$id} not found"], 404);
+        }
         $link = str_replace("storage/images/", "", $data->link);
         unlink(storage_path("app/public/images/".$link));
         $data->delete();
@@ -22,7 +25,10 @@ class ImageController extends Controller
 
     public function store(StoreUpdateImageRequest $request){
         $data = $request->validated();
-        $property = Property::findOrFail($data["property_id"]);
+        $property = Property::find($data["property_id"]);
+        if(! $property){
+            return response()->json(["message" => "Target with id={$data['property_id']} not found"], 404);
+        }
         foreach($data["files"] as $file){
             $link = $file->store("images", "public");
             Image::create(["property_id" => $property->id, "link" => "storage/".$link, "original_name" => $file->getClientOriginalName()]);
